gatherHub페이지 ssr적용완료 및 컴포넌트분리
