+++ src/components/GatherHub/JobDirectory.tsx
@@ -14,21 +14,21 @@ const JobDirectory: React.FC<JobDirectoryProps> = ({ setFilteredJob, className }
   const { isAuthenticated } = useAuth();
   const { userData } = useUserData();
 
-  // Hub 등록 여부를 useMemo로 캐싱 (불필요한 재계산 방지)
+  // Hub 등록 여부를 useMemo로 캐싱 
   const isHubRegistered = useMemo(() => userData?.hubCard || false, [userData]);
 
   // 직업군 리스트
   const jobCategories = useMemo(() => [
-    { name: '전체보기', value: 'all' },
-    { name: '프론트엔드', value: '프론트엔드' },
-    { name: '백엔드', value: '백엔드' },
-    { name: 'IOS', value: 'IOS' },
-    { name: '안드로이드', value: '안드로이드' },
-    { name: '데브옵스', value: '데브옵스' },
-    { name: '디자인', value: '디자인' },
-    { name: 'PM', value: 'PM' },
-    { name: '기획', value: '기획' },
-    { name: '마케팅', value: '마케팅' }
+    { name: '전체보기', value: 'all', hoverClass: 'hover:bg-primary hover:text-black' },
+    { name: '프론트엔드', value: '프론트엔드', hoverClass: 'hover:bg-primaryStrong hover:text-black' },
+    { name: '백엔드', value: '백엔드', hoverClass: 'hover:bg-accentOrange hover:text-black' },
+    { name: 'IOS', value: 'IOS', hoverClass: 'hover:bg-accentMaya hover:text-black' },
+    { name: '안드로이드', value: '안드로이드', hoverClass: 'hover:bg-accentPurple hover:text-black' },
+    { name: '데브옵스', value: '데브옵스', hoverClass: 'hover:bg-accentRed hover:text-black' },
+    { name: '디자인', value: '디자인', hoverClass: 'hover:bg-accentMint hover:text-black' },
+    { name: 'PM', value: 'PM', hoverClass: 'hover:bg-accentColumbia hover:text-black' },
+    { name: '기획', value: '기획', hoverClass: 'hover:bg-accentPink hover:text-black' },
+    { name: '마케팅', value: '마케팅', hoverClass: 'hover:bg-accentYellow hover:text-black' }
   ], []);
 
   // 직업군 선택 핸들러
@@ -48,7 +48,7 @@ const JobDirectory: React.FC<JobDirectoryProps> = ({ setFilteredJob, className }
   const closeLoginModal = () => setIsModalOpen(false);
 
   return (
-    <aside className={`${className} p-1 rounded-lg sticky top-4`}>
+     <aside className={`${className} p-1 rounded-lg sticky top-5 user-select-none`} style={{ userSelect: 'none' }}>
       {/* 직업 필터링 UI */}
       <JobFilter selectedJob={selectedJob} handleSelectJob={handleSelectJob} jobCategories={jobCategories} />
