+++ src/components/GatherHub/JobFilter.tsx
@@ -2,17 +2,20 @@ import React from 'react';
 import { JobFilterProps } from '@/lib/gatherHub';
 
 const JobFilter: React.FC<JobFilterProps> = ({ selectedJob, handleSelectJob, jobCategories }) => {
+  
   return (
     <>
       {/* 큰 화면용 리스트 */}
-      <ul className="hidden lg:block job-list flex-col gap-1 p-5 space-y-2 shadow mt-6 mb-6 w-[120px] h-[445px]">
+      <ul className="hidden lg:block job-list flex-col pt-4 shadow mb-6 w-[120px] h-[470px] rounded-[30px] bg-fillStrong">
         {jobCategories.map((job) => (
           <li
             key={job.value}
-            className={`job-item ${selectedJob === job.value ? 'bg-background text-primary font-bold' : 'text-gray-400'}`}
+            className={`job-item flex items-center gap-2 justify-center cursor-pointer rounded-lg p-2 transition-all duration-300 
+            ${selectedJob === job.value ? 'bg-background text-primary font-bold' : 'text-gray-400'} 
+            hover:bg-background hover:text-primary first:mt-2 last:mb-2`}
             onClick={() => handleSelectJob(job.value)}
           >
-            {selectedJob === job.value && <span>▶</span>}
+            {selectedJob === job.value && <span className="text-primary">▶</span>}
             {job.name}
           </li>
         ))}
@@ -21,7 +24,7 @@ const JobFilter: React.FC<JobFilterProps> = ({ selectedJob, handleSelectJob, job
       {/* 작은 화면용 드롭다운 */}
       <div className="block lg:hidden">
         <select
-          className="p-2 text-xl bg-black text-white rounded-lg w-full border border-gray-500"
+          className="p-2 text-xl bg-black text-white rounded-lg w-full border border-gray-500 transition-all duration-300 focus:border-blue-500 focus:bg-gray-800 hover:bg-gray-900"
           value={selectedJob}
           onChange={(e) => handleSelectJob(e.target.value)}
         >
